Prohibit participation on closed survey
